refactor(reader): split BibleReaderView into ReaderContent + ReaderMainScroller (#141)
Reshapes BibleReaderView to match main's #122 lazy-VM structure (#8dad856):
the public view becomes a thin wrapper that constructs the view model once
in .task and forwards init params to a private ReaderContent struct
defined in the same file.

Adds ReaderMainScroller.swift to encapsulate audio-scroll bookkeeping
(verseAnchors, lastScrolledVerse, applyAudioScrollIfNeeded) so the audio
feature lives in its own file rather than threading state through the
top-level view.

This reduces the shared surface area with main so future merges into the
long-running feat/bible-loop branch encounter smaller conflict zones.
